+++ typedefs.js
@@ -5,6 +5,7 @@ const { typeDefs: Drivers } = require("./data/Drivers/typeDefs");
 const { typeDefs: Status } = require("./data/Status/typeDefs");
 const { typeDefs: Results } = require("./data/Results/typeDefs");
 const { typeDefs: ConstructorStandings } = require("./data/ConstructorStandings/typeDefs");
+const { typeDefs: Qualifying } = require("./data/Qualifying/typeDefs");
 
 const base = gql`
   type PageInfo {
@@ -70,40 +71,6 @@ const base = gql`
   type DriversReport {
     nodes: [Driver]
     pageInfo: PageInfo!
-  }
-
-  "Competitor in a qualifying event"
-  type QualiCompetitor {
-    number: String
-    position: String
-    id: ID
-    driver: Driver
-    constructor: Constructor
-    "Q1 time. e.g. '1:26.736'"
-    Q1: String
-    "Q2 time. e.g. '1:26.736'"
-    Q2: String
-    "Q3 time. e.g. '1:26.736'"
-    Q3: String
-  }
-
-  type Qualifying {
-    "Formula 1"
-    series: RacingSeries
-    "2021, 2008"
-    season: String
-    "1, 2"
-    round: String
-    "Bahrain Grand Prix"
-    raceName: String
-    circuit: Circuit
-    "2021-03-28"
-    date: String
-    "15:00:00Z"
-    time: String
-    results: [QualiCompetitor]
-    "Wikipedia URL"
-    url: String
   }
 
   input ReportInput {
@@ -190,6 +157,6 @@ const base = gql`
   }
 `;
 
-const typeDefs = [base, Constructors, Drivers, Results, Status, ConstructorStandings];
+const typeDefs = [base, Constructors, Drivers, Results, Status, ConstructorStandings, Qualifying];
 
 module.exports = { typeDefs };
